refactor: Board Query, Command 서비스 분리
