+++ src/test/java/com/market/community/application/board/BoardServiceTest.java
@@ -1,9 +1,7 @@
 package com.market.community.application.board;
 
 import com.market.community.application.board.dto.BoardCreateRequest;
-import com.market.community.application.board.dto.BoardFoundResponse;
 import com.market.community.application.board.dto.BoardUpdateRequest;
-import com.market.community.application.board.dto.BoardsSimpleResponse;
 import com.market.community.domain.board.Board;
 import com.market.community.domain.board.BoardRepository;
 import com.market.community.domain.board.ImageConverter;
@@ -16,7 +14,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.mock.web.MockMultipartFile;
 
 import java.util.ArrayList;
@@ -61,39 +58,6 @@ class BoardServiceTest {
         assertThat(result).isEqualTo(1L);
     }
 
-    @Test
-    void 게시글을_모두_조회한다() {
-        // when
-        BoardsSimpleResponse result = boardService.findAllBoards(PageRequest.of(0, 10));
-
-        // then
-        assertThat(result.nextPage()).isEqualTo(-1);
-     }
-
-    @Test
-    void 게시글을_찾는다() {
-        // given
-        Board savedBoard = boardRepository.save(게시글_생성_사진없음());
-
-        // when
-        BoardFoundResponse result = boardService.findBoardById(savedBoard.getId(), 1L);
-
-        // then
-        assertSoftly(softly -> {
-            softly.assertThat(result.id()).isEqualTo(savedBoard.getId());
-            softly.assertThat(result.title()).isEqualTo(savedBoard.getPost().getTitle());
-            softly.assertThat(result.content()).isEqualTo(savedBoard.getPost().getContent());
-            softly.assertThat(result.likeCount()).isEqualTo(savedBoard.getLikeCount().getLikeCount());
-            softly.assertThat(result.isMyPost()).isEqualTo(true);
-        });
-    }
-
-    @Test
-    void 게시글이_없으면_에외를_발생시킨다() {
-        // when & then
-        assertThatThrownBy(() -> boardService.findBoardById(1L, 1L))
-                .isInstanceOf(BoardNotFoundException.class);
-    }
 
     @Test
     void 게시글을_수정한다() {
